wp-automation-master/requirements.txt
matplotlib
altair==4.2.2
numpy
openpyxl
pandas
argon2-cffi
requests
streamlit
streamlit_authenticator==0.1.5
streamlit-lottie
streamlit_option_menu
tqdm
plotly
streamlit-aggrid==0.2.3
streamlit_extras
streamlit-chat
streamlit-pandas-profiling
barfi
httpx-oauth
psycopg2-binary
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyYAML
playwright==1.41.1
st-gsheets-connection
openai